feat(snap): add content interface for device configuration
The `device-config` content interface allows another snap to seed this device
snap with initial configuration and device profiles. To use, create a new
snap with the following in snapcraft.yaml:

```
slots:
  device-config:
    interface: content
    read: $SNAP_DATA/device-profiles
```

where device-profiles is the name of a directory with the new configuration
for the device snap. You can then connect the plug in the device snap to
the slot in your snap, which will replace the configuration in the device snap.
Do this with:

```
sudo snap connect edgex-device-grove:device-config your-snap:device-config
```

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>